add performance tests for entry set functions
